+++ packages/@expo/cli/src/export/exportStaticAsync.ts
@@ -218,7 +218,8 @@ export async function exportFromServerAsync(
   Log.log(logOutput);
 
   const platform = 'web';
-  const isExportingWithSSR = exportServer && !devServer.isReactServerComponentsEnabled;
+  const isExportingWithSSR =
+    exp?.web?.output === 'server' && !devServer.isReactServerComponentsEnabled;
   const appDir = path.join(projectRoot, routerRoot);
   const faviconAsset = await generateFaviconAssetAsync(projectRoot, {
     outputDir,
@@ -735,7 +736,7 @@ function warnPossibleInvalidExportType(appDir: string, mode: Options['mode']) {
   if (apiRoutes.length) {
     // TODO: Allow API Routes for native-only.
     Log.warn(
-      chalk.yellow`Skipping export for API routes because \`web.output\` is not "server". You may want to remove the routes: ${apiRoutes
+      chalk.yellow`Skipping export for API routes because API routes are disabled. Set \`apiRoutes: true\` in the \`expo-router\` config plugin to enable them. You may want to remove the routes: ${apiRoutes
         .map((v) => path.relative(appDir, v))
         .join(', ')}`
     );
@@ -744,7 +745,7 @@ function warnPossibleInvalidExportType(appDir: string, mode: Options['mode']) {
   const middlewareFile = getMiddlewareForDirectory(appDir, mode);
   if (middlewareFile) {
     Log.warn(
-      chalk.yellow`Skipping export for middleware because \`web.output\` is not "server". You may want to remove ${path.relative(appDir, middlewareFile)}`
+      chalk.yellow`Skipping export for middleware because \`web.output\` is not "server" and API routes are disabled. Set \`apiRoutes: true\` in the \`expo-router\` config plugin to enable them. You may want to remove ${path.relative(appDir, middlewareFile)}`
     );
   }
 }