+++ packages/@expo/cli/CHANGELOG.md
@@ -7,6 +7,9 @@
 ### 🎉 New features
 
 - Enable data loaders by default for static and server output ([#50118](https://github.com/expo/expo/pull/50118) by [@hassankhan](https://github.com/hassankhan))
+
+- Add the `apiRoutes` config plugin option to support API routes with static rendering.
+
 - Enable server rendering by default for `web.output: "server"` ([#50120](https://github.com/expo/expo/pull/50120) by [@hassankhan](https://github.com/hassankhan))
 
 ### 🐛 Bug fixes